requirements.txt
    appnope
asgiref~=3.5.0
asn1crypto
attrs~=19.3.0
autopep8
backcall~=0.1.0
bleach~=3.1.0
certifi~=2019.3.9
cffi~=1.12.3
chardet~=3.0.4
cryptography~=2.7
decorator~=4.4.1
defusedxml~=0.6.0
Django~=3.2.13
django-cors-headers
django-dotenv
djangorestframework~=3.13.1
djangorestframework-simplejwt
entrypoints~=0.3
future~=0.17.1
idna~=2.8
importlib-metadata
ipykernel~=5.1.3
ipython~=7.11.1
ipython-genutils
jedi~=0.15.2
Jinja2~=2.10.3
jsonschema~=3.2.0
jupyter-client
jupyter-core
MarkupSafe~=1.1.1
mistune~=0.8.4
mongo
more-itertools
nbconvert~=5.6.1
nbformat~=5.0.3
notebook~=6.0.2
numpy~=1.18.2
opencv-python
pandocfilters~=1.4.2
parso~=0.5.2
pexpect~=4.7.0
pickleshare~=0.7.5
pika~=1.1.0
pipenv~=2018.11.26
prometheus-client
prompt-toolkit
psycopg2-binary
ptyprocess~=0.6.0
pycodestyle~=2.6.0
pycparser~=2.19
Pygments~=2.5.2
PyJWT~=2.3.0
pymongo~=3.10.1
pyrsistent~=0.15.7
pyTelegramBotAPI
python-dateutil~=2.8.1
python-telegram-bot
pytz~=2022.1
pyzmq~=18.1.1
requests~=2.23.0
scipy~=1.4.1
selenium~=3.141.0
Send2Trash~=1.5.0
six~=1.12.0
sqlparse~=0.4.2
terminado~=0.8.3
testpath~=0.4.4
toml~=0.10.2
tornado~=6.0.3
traitlets~=4.3.3
typing_extensions
urllib3~=1.25.9
virtualenv~=16.6.0
virtualenv-clone
wcwidth~=0.1.8
webencodings~=0.5.1
zipp~=1.0.0
pip~=19.3.1
colorama~=0.4.4
setuptools~=41.2.0
distlib~=0.3.4
redis~=3.3.11
tzlocal~=4.2
fastjsonschema~=2.15.3
APScheduler~=3.9.1
cachetools~=3.1.1
filelock~=3.7.0
platformdirs~=2.5.2
uritemplate~=3.0.0
google-api-python-client~=1.7.11